requirements.txt
flask
requests
pyModbusTCP
pymysql